+++ RELEASES.md
@@ -7,6 +7,7 @@ Version 0.18.0
 This version adds:
 
 * Anonymous functions (in closure syntax).  Simplifies creation of ad hoc functions.
+* Currying of function pointers.
 
 New features
 ------------
@@ -15,6 +16,8 @@ New features
 * Disallow many keywords as variables, such as `print`, `eval`, `call`, `this` etc.
 * `x.call(f, ...)` allows binding `x` to `this` for the function referenced by the function pointer `f`.
 * Anonymous functions in the syntax of a closure, e.g. `|x, y, z| x + y - z`.
+* Custom syntax now works even without the `internals` feature.
+* Currying of function pointers is supported via the `curry` keyword.
 
 Breaking changes
 ----------------
@@ -48,7 +51,7 @@ Breaking changes
 New features
 ------------
 
-* New `serde` feature to allow serializating/deserializating to/from `Dynamic` values using [`serde`](https://crates.io/crates/serde).
+* New `serde` feature to allow serializing/deserializing to/from `Dynamic` values using [`serde`](https://crates.io/crates/serde).
   This is particularly useful when converting a Rust `struct` to a `Dynamic` _object map_ and back.
 * `Engine::disable_symbol` to surgically disable keywords and/or operators.
 * `Engine::register_custom_operator` to define a custom operator.